+++ R/file.R
@@ -198,3 +198,30 @@ file_move <- function(path, new_path) {
 
   invisible(path_tidy(new))
 }
+
+#' Change file access and modification times
+#'
+#' Unlike the touch POSIX utility this does not create the file if it does not
+#' exist. Use [file_create()] to do this if needed.
+#'
+#' @template fs
+#'
+#' @examples
+#' \dontshow{.old_wd <- setwd(tempdir())}
+#' file_create("foo")
+#' file_touch("foo", "2018-01-01")
+#' file_info("foo")[c("access_time", "modification_time", "change_time", "birth_time")]
+#' \dontshow{setwd(.old_wd)}
+#' @export
+file_touch <- function(path, access_time = Sys.time(), modification_time = access_time) {
+  assert_no_missing(path)
+
+  access_time <- as.POSIXct(access_time)
+  modification_time <- as.POSIXct(modification_time)
+
+  path <- path_expand(path)
+
+  touch_(path, access_time, modification_time)
+
+  invisible(path_tidy(path))
+}