+++ src/RcppExports.cpp
@@ -131,6 +131,18 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
+// touch_
+void touch_(CharacterVector path, double atime, double mtime);
+RcppExport SEXP _fs_touch_(SEXP pathSEXP, SEXP atimeSEXP, SEXP mtimeSEXP) {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< CharacterVector >::type path(pathSEXP);
+    Rcpp::traits::input_parameter< double >::type atime(atimeSEXP);
+    Rcpp::traits::input_parameter< double >::type mtime(mtimeSEXP);
+    touch_(path, atime, mtime);
+    return R_NilValue;
+END_RCPP
+}
 // cleanup_
 void cleanup_();
 RcppExport SEXP _fs_cleanup_() {
@@ -309,6 +321,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_fs_unlink_", (DL_FUNC) &_fs_unlink_, 1},
     {"_fs_copyfile_", (DL_FUNC) &_fs_copyfile_, 3},
     {"_fs_chown_", (DL_FUNC) &_fs_chown_, 3},
+    {"_fs_touch_", (DL_FUNC) &_fs_touch_, 3},
     {"_fs_cleanup_", (DL_FUNC) &_fs_cleanup_, 0},
     {"_fs_getmode_", (DL_FUNC) &_fs_getmode_, 2},
     {"_fs_strmode_", (DL_FUNC) &_fs_strmode_, 1},